add animated favorites menu, filmCardComponent to render cards
