+++ .idea/workspace.xml
@@ -5,11 +5,24 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="154ebf74-2b4c-4c71-a8b2-acfb77888ecd" name="Changes" comment="">
+      <change afterPath="$PROJECT_DIR$/src/core/components/filmCardComponent.ts" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/src/style/pages/favoritesOverlay.scss" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/package-lock.json" beforeDir="false" afterPath="$PROJECT_DIR$/package-lock.json" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/components/filmListComponent.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/components/filmListComponent.ts" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/core/components/filmSliderComponent.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/components/filmSliderComponent.ts" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/core/components/searchLineComponent.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/components/searchLineComponent.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/constants/SectionID.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/constants/SectionID.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/router.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/router.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/core/view.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/core/view.ts" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/style/common/filmCard.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/common/filmCard.scss" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/style/common/slider.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/common/slider.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/style/components/footer.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/components/footer.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/style/components/header.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/components/header.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/style/index.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/index.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/style/pages/mainPage.scss" beforeDir="false" afterPath="$PROJECT_DIR$/src/style/pages/mainPage.scss" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/favoritesView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/favoritesView.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/filmsListView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/filmsListView.ts" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/views/responseErrorView.ts" beforeDir="false" afterPath="$PROJECT_DIR$/src/views/responseErrorView.ts" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -118,6 +131,7 @@
       <workItem from="1638164170729" duration="765000" />
       <workItem from="1638165505992" duration="21387000" />
       <workItem from="1638349814522" duration="4396000" />
+      <workItem from="1638367670266" duration="7076000" />
     </task>
     <servers />
   </component>
