+++ src/style/common/slider.scss
@@ -1,3 +1,4 @@
+//it's a bad practise to use magic values that can be changed
 .swiper {
   width: 1000px;
   height: 450px;
