+++ src/style/components/footer.scss
@@ -12,3 +12,5 @@
 #gitLabel {
   color: $text;
 }
+
+//styling which is based on id's not the best practise, it's difficult to override sometimes
