+++ src/style/components/header.scss
@@ -22,6 +22,7 @@ $favoriteButtonSize: 40px;
 #switchFavorites {
   height: $favoriteButtonSize;
   width: $favoriteButtonSize;
+  transition: all .2s ease-in-out;
 }
 
 #switchFavorites:hover {
