+++ src/style/index.scss
@@ -2,6 +2,7 @@
 @import 'global/global';
 @import 'components/header';
 @import 'pages/mainPage';
+@import 'pages/favoritesOverlay';
 @import 'components/footer';
 @import 'common/filmCard';
 @import 'common/slider';
