+++ src/style/pages/mainPage.scss
@@ -1,4 +1,5 @@
 #content {
+  position: relative;
   background: $backgroundLight;
   height: 85%;
   display: flex;
@@ -14,11 +15,12 @@
 }
 
 .search-form{
-    display: flex;
-    width: 100%;
-    background-color: rgba(242,221,146,0.9);
-    border-radius: 30px;
-    padding: 7px 5px 7px 53px;
+  display: flex;
+  width: 100%;
+  background-color: rgba(242,221,146,0.9);
+  border-radius: 30px;
+  //what the hell is this values? XD
+  padding: 7px 5px 7px 53px;
 }
 
 .search-input {
@@ -26,6 +28,10 @@
   border: none;
   outline: none;
   background: none;
+}
+
+.search-submit {
+  transition: all .2s ease-in-out;
 }
 
 .search-submit:hover {
