+++ src/views/favoritesView.ts
@@ -23,5 +23,13 @@ export default class FavoritesView extends View<{
     });
 
     this.container.append(favoritesTitle, favoritesList);
+    const overlay = document.querySelector(".overlay") as HTMLElement;
+    overlay.style.width = '50%';
+  }
+
+  public clear(): void {
+    this.container.innerHTML = '';
+    const overlay = document.querySelector(".overlay") as HTMLElement;
+    overlay.style.width = '0%';
   }
 }
